fix:看板计算公式调整

+++ HymsonMES/src/Libraries/Hymson.MES.Services/Dtos/Report/ProductionManagePanelReportDto.cs
@@ -409,7 +409,7 @@ namespace Hymson.MES.Services.Dtos.Report
         /// <summary>
         /// 不良数
         /// </summary>
-        public int UnqualifiedQty { get; set; }
+        public decimal UnqualifiedQty { get; set; }
 
         /// <summary>
         /// 不良率
